test: remove 12 duplicate root browser tests, trim IndexedDBHelper, fix timing issues
Root browser tests deleted (features/ versions are identical or better):
- 7 exact duplicates: AboutPage, CharacterCard, CharacterForm,
  CharacterSidebarItem, ChatArea, ChatList, PersonaForm
- 5 near-duplicates merged into features/: CharacterList, CharacterSidebar,
  GenerationParameters, PersonaList, SettingsModal
- character.test.ts and ChatPersion.test.ts (features/pages/integration versions better)

IndexedDBHelper.test.ts: trimmed from 22 to 3 tests.
  Kept schema structure tests; removed Dexie CRUD tests (library guarantee,
  already covered by adapter tests).

Timing fixes:
- fontSettings.test.ts: setTimeout -> vi.waitFor
- CharacterSettingsModal.test.ts: 600ms setTimeout -> vi.waitFor

A11y fixes:
- SettingsModal: getByText -> getByRole (a11y-first)
- CharacterSidebar: querySelector -> getByRole (a11y-first)
