--- frontend/test/browser/CharacterForm.test.ts
@@ -1,177 +1,0 @@
-/// <reference types="vitest/browser" />
-import { test, expect, describe, vi, beforeEach } from "vitest";
-import { render } from "vitest-browser-svelte";
-import CharacterForm from "@/features/character/components/CharacterForm.svelte";
-import { characterStore } from "@/features/character/stores/characterStore.svelte";
-import type { Character } from "@arisutalk/character-spec/v0/Character";
-
-// Mock the characterStore
-vi.mock("@/features/character/stores/characterStore.svelte", () => {
-    const mockStore = {
-        characters: [],
-        importCharacter: vi.fn(),
-        add: vi.fn(),
-        delete: vi.fn(),
-    };
-    return {
-        characterStore: mockStore,
-        CharacterStore: vi.fn(),
-    };
-});
-
-describe("CharacterForm Component", () => {
-    let mockCharacter: Character;
-    const mockOnSave = vi.fn() satisfies () => void;
-    const mockOnCancel = vi.fn() satisfies () => void;
-    const mockOnSubmit = vi.fn() satisfies (character: Character) => void;
-
-    beforeEach(() => {
-        vi.clearAllMocks();
-        mockOnSave.mockClear();
-        mockOnCancel.mockClear();
-        mockOnSubmit.mockClear();
-
-        mockCharacter = {
-            id: "char-1",
-            specVersion: 0,
-            name: "Test Character",
-            description: "A test character",
-            avatarUrl: "https://example.com/avatar.png",
-            assets: { assets: [] },
-            prompt: {
-                description: "",
-                authorsNote: "",
-                lorebook: { config: { tokenLimit: 0 }, data: [] },
-            },
-            executables: {
-                runtimeSetting: { mem: undefined, timeout: 30000 },
-                replaceHooks: { display: [], input: [], output: [], request: [] },
-            },
-            metadata: {
-                author: undefined,
-                license: "",
-                version: undefined,
-                distributedOn: undefined,
-                additionalInfo: undefined,
-            },
-        };
-
-        characterStore.characters = [];
-        characterStore.importCharacter = vi.fn();
-    });
-
-    test("renders create form when no character is provided", async () => {
-        const { getByLabelText } = render(CharacterForm, {
-            onSave: mockOnSave,
-            onCancel: mockOnCancel,
-        });
-
-        await expect.element(getByLabelText(/Name/i)).toBeVisible();
-        await expect.element(getByLabelText(/Description/i)).toBeVisible();
-    });
-
-    test("renders edit form when character is provided", async () => {
-        const { getByLabelText } = render(CharacterForm, {
-            character: mockCharacter,
-            onSave: mockOnSave,
-            onCancel: mockOnCancel,
-        });
-
-        await expect.element(getByLabelText(/Name/i)).toHaveValue("Test Character");
-        await expect.element(getByLabelText(/Description/i)).toHaveValue("A test character");
-    });
-
-    test("calls onSubmit when form is submitted", async () => {
-        const { getByRole } = render(CharacterForm, {
-            character: mockCharacter,
-            onSubmit: mockOnSubmit,
-            onSave: mockOnSave,
-        });
-
-        const submitButton = getByRole("button", { name: /Save/i });
-        await submitButton.click();
-
-        expect(mockOnSubmit).toHaveBeenCalledWith(
-            expect.objectContaining({
-                name: "Test Character",
-                description: "A test character",
-            })
-        );
-    });
-
-    test("shows error message when name is empty", async () => {
-        const { getByLabelText, getByText, container } = render(CharacterForm, {
-            onCancel: mockOnCancel,
-        });
-
-        const nameInput = getByLabelText(/Name/i);
-        await nameInput.fill(""); // Clear name
-
-        const submitButton = getByText(/Save/i);
-        await submitButton.click();
-
-        const errorMessage = container.querySelector(".alert-error");
-        expect(errorMessage).toBeTruthy();
-        expect(errorMessage?.textContent).toContain("Name is required");
-    });
-
-    test("calls onCancel when cancel button is clicked", async () => {
-        const { getByRole } = render(CharacterForm, {
-            onCancel: mockOnCancel,
-        });
-
-        // Click cancel button in the active tab (Create New)
-        const cancelButton = getByRole("button", { name: /Cancel/i }).first();
-        await cancelButton.click();
-
-        expect(mockOnCancel).toHaveBeenCalledOnce();
-    });
-
-    test("has file upload input for importing", async () => {
-        const { getByLabelText } = render(CharacterForm, {
-            onSave: mockOnSave,
-            onCancel: mockOnCancel,
-        });
-
-        const fileInput = getByLabelText(/Import/i);
-        await expect.element(fileInput).toBeVisible();
-    });
-
-    test("handles character import", async () => {
-        characterStore.importCharacter = vi.fn().mockResolvedValue({
-            success: true,
-            character: mockCharacter,
-        });
-
-        const { getByLabelText } = render(CharacterForm, {
-            onSave: mockOnSave,
-            onCancel: mockOnCancel,
-        });
-
-        const fileInput = getByLabelText(/Import/i);
-        await fileInput.upload(new File(['{"name":"Imported Char","specVersion":0}'], "character.json", { type: "application/json" }));
-
-        await vi.waitFor(() => {
-            expect(characterStore.importCharacter).toHaveBeenCalled();
-        });
-    });
-
-    test("shows import error when import fails", async () => {
-        characterStore.importCharacter = vi.fn().mockResolvedValue({
-            success: false,
-            error: "Invalid file format",
-        });
-
-        const { getByLabelText } = render(CharacterForm, {
-            onSave: mockOnSave,
-            onCancel: mockOnCancel,
-        });
-
-        const fileInput = getByLabelText(/Import/i);
-        await fileInput.upload(new File(['{"name":"Bad Char"}'], "bad.json", { type: "application/json" }));
-
-        await vi.waitFor(() => {
-            expect(characterStore.importCharacter).toHaveBeenCalled();
-        });
-    });
-});
