--- frontend/test/browser/ChatList.test.ts
@@ -1,200 +1,0 @@
-/// <reference types="vitest/browser" />
-import { test, expect, describe, vi, beforeEach } from "vitest";
-import { render } from "vitest-browser-svelte";
-import ChatList from "@/features/chat/components/ChatList.svelte";
-import { chatStore } from "@/features/chat/stores/chatStore.svelte";
-import type { LocalChat } from "@/lib/interfaces";
-
-// Mock the chat store
-vi.mock("@/features/chat/stores/chatStore.svelte", () => {
-    const mockStore = {
-        chats: [],
-        activeChatId: null,
-        createChat: vi.fn().mockResolvedValue("new-chat-id"),
-        setActiveChat: vi.fn(),
-        deleteChat: vi.fn().mockResolvedValue(undefined),
-        isGenerating: false,
-        activeMessages: [],
-        sendMessage: vi.fn(),
-    };
-    return {
-        chatStore: mockStore,
-        ChatStore: vi.fn(),
-    };
-});
-
-describe("ChatList Component", () => {
-    beforeEach(() => {
-        vi.clearAllMocks();
-        chatStore.chats = [];
-        chatStore.activeChatId = null;
-    });
-
-    test("renders header with chat title and new chat button", async () => {
-        const { getByRole, getByLabelText } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        const heading = getByRole("heading", { name: "Chats", level: 3 });
-        await expect.element(heading).toBeVisible();
-
-        const newChatButton = getByLabelText("New Chat");
-        await expect.element(newChatButton).toBeVisible();
-    });
-
-    test("renders empty state when no chats exist", async () => {
-        const { getByText } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        const emptyMessage = getByText("No chats yet.");
-        await expect.element(emptyMessage).toBeVisible();
-
-        const createLink = getByText("Create one?");
-        await expect.element(createLink).toBeVisible();
-    });
-
-    test("displays chats filtered by characterId", async () => {
-        chatStore.chats = [
-            {
-                id: "chat-1",
-                name: "Chat 1",
-                characterId: "char-1",
-                createdAt: 0,
-                updatedAt: 0,
-                lastMessage: 0,
-                title: "Chat 1",
-            } satisfies LocalChat,
-            {
-                id: "chat-2",
-                name: "Chat 2",
-                characterId: "char-2",
-                createdAt: 0,
-                updatedAt: 0,
-                lastMessage: 0,
-                title: "Chat 2",
-            } satisfies LocalChat,
-        ];
-
-        const { getByText } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        await expect.element(getByText("Chat 1")).toBeVisible();
-    });
-
-    test("highlights active chat", async () => {
-        chatStore.chats = [
-            {
-                id: "chat-1",
-                name: "Chat 1",
-                characterId: "char-1",
-                createdAt: 0,
-                updatedAt: 0,
-                lastMessage: 0,
-                title: "Chat 1",
-            } satisfies LocalChat,
-        ];
-        chatStore.activeChatId = "chat-1";
-
-        const { getByRole } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        const chatButton = getByRole("button", { name: "Chat 1" });
-        await expect.element(chatButton).toBeVisible();
-        // The active chat should have the 'menu-active' class
-    });
-
-    test("creates new chat when new chat button is clicked", async () => {
-        const { getByLabelText } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        const newChatButton = getByLabelText("New Chat");
-        await newChatButton.click();
-
-        expect(chatStore.createChat).toHaveBeenCalledWith("char-1", "Chat 1");
-    });
-
-    test("selects chat when chat item is clicked", async () => {
-        chatStore.chats = [
-            {
-                id: "chat-1",
-                name: "Chat 1",
-                characterId: "char-1",
-                createdAt: 0,
-                updatedAt: 0,
-                lastMessage: 0,
-                title: "Chat 1",
-            } satisfies LocalChat,
-        ];
-
-        const { getByRole } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        const chatButton = getByRole("button", { name: "Chat 1" });
-        await chatButton.click();
-
-        expect(chatStore.setActiveChat).toHaveBeenCalledWith("chat-1");
-    });
-
-    test("deletes chat when delete button is clicked", async () => {
-        // Mock window.confirm
-        const originalConfirm = window.confirm;
-        window.confirm = vi.fn(() => true);
-
-        chatStore.chats = [
-            {
-                id: "chat-1",
-                name: "Chat 1",
-                characterId: "char-1",
-                createdAt: 0,
-                updatedAt: 0,
-                lastMessage: 0,
-                title: "Chat 1",
-            } satisfies LocalChat,
-        ];
-
-        const { getByRole, getByLabelText } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        // Hover over the chat item to reveal the delete button
-        const chatButton = getByRole("button", { name: "Chat 1" });
-        await chatButton.hover();
-
-        // Click the delete button
-        const deleteButton = getByLabelText("Delete");
-        await deleteButton.click();
-
-        expect(window.confirm).toHaveBeenCalled();
-        expect(chatStore.deleteChat).toHaveBeenCalledWith("chat-1");
-
-        window.confirm = originalConfirm;
-    });
-
-    test("generates sequential chat names", async () => {
-        chatStore.chats = [
-            {
-                id: "chat-1",
-                name: "Chat 1",
-                characterId: "char-1",
-                createdAt: 0,
-                updatedAt: 0,
-                lastMessage: 0,
-                title: "Chat 1",
-            } satisfies LocalChat,
-        ];
-
-        const { getByLabelText } = render(ChatList, {
-            characterId: "char-1",
-        });
-
-        const newChatButton = getByLabelText("New Chat");
-        await newChatButton.click();
-
-        expect(chatStore.createChat).toHaveBeenCalledWith("char-1", "Chat 2");
-    });
-});
