--- frontend/test/browser/ChatPersion.test.ts
@@ -1,85 +1,0 @@
-/// <reference types="vitest/browser" />
-import { test, expect, describe, vi, afterEach } from "vitest";
-import { render } from "vitest-browser-svelte";
-import CharacterLayoutTestWrapper from "./CharacterLayoutTestWrapper.svelte";
-
-describe("Persona and Chat interactions", () => {
-    afterEach(() => {
-        vi.restoreAllMocks();
-    });
-
-    test("Manage Personas opens persona modal", async () => {
-        const { getByLabelText, container } = render(CharacterLayoutTestWrapper);
-
-        const personaBtn = getByLabelText("Manage Personas");
-        await expect.element(personaBtn).toBeVisible();
-        await personaBtn.click();
-
-        // Wait for the dialog to become visible (open attribute)
-        await vi.waitFor(
-            () => {
-                const dialog = container.querySelector("#persona_modal");
-                expect(dialog?.hasAttribute("open")).toBe(true);
-            },
-            { timeout: 5000 }
-        );
-    });
-
-    // TODO: Fix flaky test - fake timers incompatible with vitest browser mode
-    // This test also fails on the kei branch. Needs a different approach
-    // for waiting on bot responses without fake timers.
-    test.skip("Create character and send chat message", async () => {
-        // Import and configure chatStore with Mock provider for testing
-        // Mock chatStore.waitForSettings to avoid delays
-
-        const { chatStore } = await import("@/features/chat/stores/chatStore.svelte");
-        // Cast chatStore to access all private members
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        vi.spyOn(chatStore as any, "waitForSettings").mockResolvedValue(undefined);
-        await chatStore.initPromise;
-        await chatStore.setProvider("MOCK", {
-            mockDelay: 50,
-            responses: ["Response 1", "Response 2"],
-            generationParameters: {},
-        });
-
-        const { getByLabelText, getByText, getByRole } = render(CharacterLayoutTestWrapper);
-
-        // Open Add Character modal
-        const addCharButton = getByLabelText("Add Character");
-        await addCharButton.click();
-
-        const nameInput = getByLabelText("Name");
-        await expect.element(nameInput).toBeVisible();
-        await nameInput.fill("TestBot");
-
-        const saveBtn = getByText("Save Character");
-        await saveBtn.click();
-
-        // Select the created character from the sidebar
-        const charBtn = getByRole("button", { name: "TestBot" });
-        await expect.element(charBtn).toBeVisible();
-        await charBtn.click();
-
-        // Create a new chat for the character so ChatArea becomes active
-        const newChatBtn = getByLabelText("New Chat");
-        await expect.element(newChatBtn).toBeVisible();
-        await newChatBtn.click();
-
-        // Send a user message (chat input uses placeholder)
-        const input = getByRole("textbox", { name: "Type a message..." });
-        await expect.element(input).toBeVisible();
-        await input.fill("Hello there");
-
-        const sendBtn = getByText("Send");
-        await sendBtn.click();
-
-        // User message appears
-        const userMsg = getByText("Hello there");
-        await expect.element(userMsg).toBeVisible();
-
-        // Wait for the mocked bot response (uses real timers in browser mode)
-        const botMsg = getByText("Response 1");
-        await expect.element(botMsg).toBeVisible();
-    });
-});
