--- frontend/test/browser/SettingsModal.test.ts
@@ -1,91 +1,0 @@
-import { test, expect, describe, vi, beforeEach } from "vitest";
-import { render } from "vitest-browser-svelte";
-import SettingsModal from "@/components/SettingsModal.svelte";
-import { settings } from "@/lib/stores/settings.svelte";
-import { SettingsSchema } from "@/lib/types/IDataModel";
-import { Logger } from "@common/logger/Logger";
-
-// Mock uiState
-vi.mock("@/lib/stores/ui.svelte", () => ({
-    uiState: {
-        settingsModalOpen: true,
-        openSettingsModal: vi.fn(),
-        closeSettingsModal: vi.fn(),
-    },
-}));
-
-describe("SettingsModal Component", () => {
-    beforeEach(() => {
-        // Reset stores
-        settings.value = SettingsSchema.parse({});
-        // We need to ensure settings.save is mockable or we just let it run (it will try to call adapter)
-        // Better to mock the adapter or the save method.
-        // Since we import the instance, we can spy on it.
-        vi.spyOn(settings, "save").mockResolvedValue(undefined);
-        vi.spyOn(Logger, "structured").mockImplementation(() => {});
-    });
-
-    test("logs modal.open on mount", async () => {
-        render(SettingsModal);
-
-        expect(Logger.structured).toHaveBeenCalledWith(
-            "modal.open",
-            expect.objectContaining({
-                modalName: "SettingsModal",
-            })
-        );
-    });
-
-    test("logs modal.close on close", async () => {
-        const { getByLabelText } = render(SettingsModal);
-
-        await getByLabelText("Close").click();
-
-        expect(Logger.structured).toHaveBeenCalledWith(
-            "modal.close",
-            expect.objectContaining({
-                modalName: "SettingsModal",
-            })
-        );
-    });
-
-    test("renders correctly", async () => {
-        const { getByRole, getByLabelText } = render(SettingsModal);
-
-        // Use getByRole for heading to be more specific
-        await expect.element(getByRole("heading", { name: "Settings", level: 2 })).toBeVisible();
-        await expect.element(getByLabelText("General Settings")).toBeVisible();
-    });
-
-    test("switches tabs", async () => {
-        const { getByRole } = render(SettingsModal);
-
-        // Initial tab is General
-        await expect.element(getByRole("heading", { name: "General Settings" })).toBeVisible();
-
-        // Switch to LLM
-        await getByRole("button", { name: "LLM Configuration" }).click();
-        await expect.element(getByRole("heading", { name: "LLM Configuration" })).toBeVisible();
-
-        // Switch to Advanced
-        await getByRole("button", { name: "Advanced Settings" }).click();
-        await expect.element(getByRole("heading", { name: "Advanced Settings" })).toBeVisible();
-        await expect.element(getByRole("combobox")).toBeVisible();
-    });
-
-    test("adds new LLM config", async () => {
-        const { getByText, getByRole } = render(SettingsModal);
-
-        await getByRole("button", { name: "LLM Configuration" }).click();
-
-        // Before: no configs
-        expect(settings.value.llmConfigs.length).toBe(0);
-
-        // Click Add Model
-        await getByText("Add Model").click();
-
-        // After: one config added
-        expect(settings.value.llmConfigs.length).toBe(1);
-        expect(settings.value.llmConfigs[0].name).toBe("Model 1");
-    });
-});
