+++ frontend/test/lib/adapters/IndexedDBHelper.test.ts
@@ -2,7 +2,6 @@ import "fake-indexeddb/auto";
 import { describe, it, expect, beforeEach } from "vitest";
 import { getArisuDB } from "@/lib/adapters/storage/IndexedDBHelper";
 import { CharacterSchema, type Character } from "@arisutalk/character-spec/v0/Character";
-import { SettingsSchema, type Settings } from "@/lib/types/IDataModel";
 import { apply } from "@arisutalk/character-spec/utils";
 
 describe("IndexedDBHelper (ArisuDB)", () => {
@@ -27,159 +26,6 @@ describe("IndexedDBHelper (ArisuDB)", () => {
 
         it("opens successfully", async () => {
             await expect(db.open()).resolves.not.toThrow();
-        });
-    });
-
-    describe("Characters table", () => {
-        const testCharacter: Character = apply(CharacterSchema, {
-            id: "char-1",
-            name: "Test Character",
-            specVersion: 0,
-            description: "A test character",
-            assets: {
-                assets: [],
-            },
-            prompt: {
-                description: "",
-            },
-            metadata: {},
-        });
-
-        it("puts and gets a character", async () => {
-            await db.characters.put(testCharacter);
-            const result = await db.characters.get("char-1");
-            expect(result?.name).toBe("Test Character");
-        });
-
-        it("returns undefined for non-existent character", async () => {
-            const result = await db.characters.get("non-existent");
-            expect(result).toBeUndefined();
-        });
-
-        it("deletes a character", async () => {
-            await db.characters.put(testCharacter);
-            await db.characters.delete("char-1");
-            const result = await db.characters.get("char-1");
-            expect(result).toBeUndefined();
-        });
-
-        it("lists all characters", async () => {
-            await db.characters.put(testCharacter);
-            await db.characters.put({ ...testCharacter, id: "char-2", name: "Second" });
-            const all = await db.characters.toArray();
-            expect(all.length).toBe(2);
-        });
-
-        it("clears all characters", async () => {
-            await db.characters.put(testCharacter);
-            await db.characters.clear();
-            const all = await db.characters.toArray();
-            expect(all.length).toBe(0);
-        });
-    });
-
-    describe("Chats table with indexes", () => {
-        const testChat = {
-            id: "chat-1",
-            characterId: "char-1",
-            title: "Test Chat",
-            createdAt: Date.now(),
-            updatedAt: Date.now(),
-        };
-
-        it("queries by characterId index", async () => {
-            await db.chats.put(testChat);
-            await db.chats.put({ ...testChat, id: "chat-2", characterId: "char-2" });
-            const results = await db.chats.where("characterId").equals("char-1").toArray();
-            expect(results.length).toBe(1);
-            expect(results[0].id).toBe("chat-1");
-        });
-
-        it.concurrent("orders by createdAt", async () => {
-            const now = Date.now();
-            await db.chats.put({ ...testChat, id: "chat-old", createdAt: now - 1000 });
-            await db.chats.put({ ...testChat, id: "chat-new", createdAt: now });
-            const results = await db.chats.orderBy("createdAt").toArray();
-            expect(results[0].id).toBe("chat-old");
-            expect(results[1].id).toBe("chat-new");
-        });
-    });
-
-    describe("Messages table with indexes", () => {
-        const testMessage = {
-            id: "msg-1",
-            chatId: "chat-1",
-            role: "user" as const,
-            content: { type: "text" as const, data: "Hello" },
-            timestamp: Date.now(),
-            inlays: [],
-        };
-
-        it("queries by chatId index", async () => {
-            await db.messages.put(testMessage);
-            await db.messages.put({ ...testMessage, id: "msg-2", chatId: "chat-2" });
-            const results = await db.messages.where("chatId").equals("chat-1").toArray();
-            expect(results.length).toBe(1);
-        });
-
-        it("queries by role", async () => {
-            await db.messages.put(testMessage);
-            await db.messages.put({ ...testMessage, id: "msg-2", role: "assistant" });
-            const users = await db.messages.where("role").equals("user").toArray();
-            expect(users.length).toBe(1);
-        });
-
-        it("handles bulk put", async () => {
-            const messages = [
-                { ...testMessage, id: "msg-1" },
-                { ...testMessage, id: "msg-2" },
-                { ...testMessage, id: "msg-3" },
-            ];
-            await db.messages.bulkPut(messages);
-            const all = await db.messages.toArray();
-            expect(all.length).toBe(3);
-        });
-    });
-
-    describe("Personas table", () => {
-        const testPersona = {
-            id: "persona-1",
-            name: "Test Persona",
-            description: "A test persona",
-            assets: { assets: [] },
-        };
-
-        it("puts and gets a persona", async () => {
-            await db.personas.put(testPersona);
-            const result = await db.personas.get("persona-1");
-            expect(result?.name).toBe("Test Persona");
-        });
-
-        it("queries by name index", async () => {
-            await db.personas.put(testPersona);
-            await db.personas.put({ ...testPersona, id: "persona-2", name: "Another" });
-            const results = await db.personas.where("name").equals("Test Persona").toArray();
-            expect(results.length).toBe(1);
-        });
-    });
-
-    describe("Settings table", () => {
-        const testSettings: Settings & { id: string } = {
-            ...SettingsSchema.parse({}),
-            id: "singleton",
-        };
-
-        it("stores singleton settings", async () => {
-            await db.settings.put(testSettings);
-            const result = await db.settings.get("singleton");
-            expect(result?.theme).toBe("system");
-        });
-
-        it("updates existing settings", async () => {
-            await db.settings.put(testSettings);
-            await db.settings.put({ ...testSettings, theme: "light" });
-            const result = await db.settings.get("singleton");
-            expect(result?.theme).toBe("light");
         });
     });
